requirements.txt
requests
websockets
pyyaml